// src/types/processedInvestor.ts
export type ProcessedInvestorData = {
  id: number;
  limited_partner_name: string;
  limited_partner_former_name: string | null;
  limited_partner_also_known_as: string | null;
  limited_partner_type: string | null;
  aum: number | null;
  year_founded: number | null;
  website: string | null;
  hqlocation: string | null;
  hqaddress_line1: string | null;
  hqaddress_line2: string | null;
  hqcity: string | null;
  hqstate_province: string | null;
  hqpost_code: string | null;
  hqcountry: string | null;
  hqphone: string | null;
  hqemail: string | null;
  description: string | null;
  primary_contact_pbid: string | null;
  primary_contact: string | null;
  primary_contact_title: string | null;
  primary_contact_email: string | null;
  primary_contact_phone: string | null;
  total_commitments_in_debt_funds: number | null;
  total_commitments_in_pefunds: number | null;
  total_commitments_in_refunds: number | null;
  total_commitments_in_vcfunds: number | null;
  total_commitments_in_fofs_and2nd: number | null;
  total_commitments_in_infrastructure: number | null;
  total_commitments_in_energy_funds: number | null;
  total_commitments_in_other_funds: number | null;
  direct_investments: number | null;
  open_to_first_time_funds: string | null;
  number_of_affiliated_funds: number | null;
  number_of_affiliated_investors: number | null;
  allocation_to_alternative_investments: number | null;
  allocation_to_alternative_investments_percent: number | null;
  private_equity: number | null;
  private_equity_percent: number | null;
  real_estate: number | null;
  real_estate_percent: number | null;
  special_opportunities: number | null;
  special_opportunities_percent: number | null;
  hedge_funds: number | null;
  hedge_funds_percent: number | null;
  equities: number | null;
  equities_percent: number | null;
  fixed_income: number | null;
  fixed_income_percent: number | null;
  cash: number | null;
  cash_percent: number | null;
  policy_description: string | null;
  preferred_commitment_size_min: number | null;
  preferred_commitment_size_max: number | null;
  preferred_direct_investment_size_min: number | null;
  preferred_direct_investment_size_max: number | null;
  preferred_geography: string | null;
  preferred_fund_type: string | null;
  target_alternatives_min: number | null;
  target_alternatives_max: number | null;
  target_private_equity_min: number | null;
  target_private_equity_max: number | null;
  target_real_estate_min: number | null;
  target_real_estate_max: number | null;
  target_special_opportunities_min: number | null;
  target_special_opportunities_max: number | null;
};